Major refactor - Navbar navigation in LandingPage.tsx
