+++ frontend/src/screens/MenuPage.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function MenuPage() {
+    
+  return (
+    
+    <div>MenuPage</div>
+  )
+}
